refactor(#1) : 플레이어 쿼리키 정리, 불필요한 파일 삭제
