+++ src/components/mypage/WriteList.tsx
@@ -7,7 +7,7 @@ import BoardItem from './BoardItem'
 import BoardNoData from './BoardNoData'
 import { getCurrentMusicData, updateCurrentMusic } from '@/shared/main/api'
 import { useSession } from 'next-auth/react'
-import { GET_MUSICLIST_QUERY_KEY } from '@/query/musicPlayer/musicPlayerQueryKey'
+import { GET_MUSICLIST_QUERY_KEY } from '@/query/musicPlayer/musicPlayerQueryKeys'
 import { paging } from '@/util/util'
 import Pagination from '@/util/Pagination '
 import type { Board } from '@/types/mypage/types'
