+++ src/components/userpage/WriteList.tsx
@@ -8,7 +8,7 @@ import BoardNoData from '../mypage/BoardNoData'
 import { getUserVisibilityData } from '@/shared/userpage/api'
 import LockContents from './LockContents'
 import { useSession } from 'next-auth/react'
-import { GET_MUSICLIST_QUERY_KEY } from '@/query/musicPlayer/musicPlayerQueryKey'
+import { GET_MUSICLIST_QUERY_KEY } from '@/query/musicPlayer/musicPlayerQueryKeys'
 import { paging } from '@/util/util'
 import Pagination from '@/util/Pagination '
 import type { Board } from '@/types/mypage/types'
