--- src/query/queryKey.ts
@@ -1,3 +1,0 @@
-export const QUERY_KEY = {
-  TEST: 'test',
-}
